requirements.txt
cassandra-driver
cryptography
boto3
fastapi
gunicorn
jupyter
pandas
pypyr
python-dotenv
tensorflow
uvicorn